add basic test structure
