+++ server/finatra-server/src/main/scala/tapir/server/finatra/FinatraRoute.scala
@@ -1,0 +1,5 @@
+package tapir.server.finatra
+import com.twitter.finagle.http.{Request, Response}
+import com.twitter.util.Future
+
+case class FinatraRoute(handler: Request => Future[Response], path: String)
